MovingPlatforms:
I ahve added a platform scene. I have made a script with a parent function so that the player moves with the platform. There are two box colliders, one of them is is trigger on top of the platform (where the player will stand). Apart from the script, you just need an animation to make the platforms move.